Fix Tuya BlitzWolf motion sensor (#3629)
* Fix Tuya BlitzWolf motion sensor using wrong cluster

* Add custom `TuyaMotionWithReset` cluster

This code is somewhat copied from `MotionWithReset` and related clusters from the base `zhaquirks` module.
However, that code is outdated and still uses `Bus` somewhat and listens to the IAS zone status change commands. We generally listen to IAS zone status attribute updates now. This is also needed for the Tuya DP conversion.

* Expand test to except automatic reset