IKEA Vallhorn quirk
Allows to configure the "on only when dark" and the "on time"
value. Both "only when dark" (on/off) and "on time" (only 60s or 300s) can also be
configured by buttons on the device itself.

There is another attribute 0x0004 in this cluster which I haven't figured out what
it does.